do not retry a close() call after error
Modern man pages say:

"... a failure return [of close] should be used only for diagnostic
purposes (i.e., a warning to the application that there may still be
I/O pending or there may have been failed I/O) or remedial purposes
(e.g., writing the file once more or creating a backup).

Retrying the close() after a failure return is the wrong thing to
do, since this may cause a reused file descriptor from another thread
to be closed. This can occur because the Linux kernel always releases
the file descriptor early in the close operation, freeing it for reuse;
the steps that may return an error, such as flushing data to the filesystem
or device, occur only later in the close operation.

Many other implementations similarly always close the file descriptor
(except in the case of EBADF, meaning that the file descriptor was invalid)
even if they subsequently report an error on return from close().

POSIX.1 is currently silent on this point, but there are plans to mandate
this behavior in the next major release of the standard."